Gran Repartición del Visor Dividido las diferentes partes del visor según si es una parte o se usa para posicionarlos en el mostrador.
