+++ func/visor/partes/regla.js
@@ -1,0 +1,3 @@
+export function cargar_regla() {
+	return crear_div("regla");
+}
